backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
langchain==0.3.12
langchain-community==0.3.12
langchain-openai==0.2.14
chromadb==0.5.23
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
requests==2.32.3
python-multipart==0.0.17
websockets==14.0
matplotlib==3.10.0
plotly==5.24.1
pandas==2.2.3
numpy==2.2.1
python-dotenv==1.0.1
aiofiles==24.1.0
aiosqlite==0.20.0
email-validator==2.2.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
jieba==0.42.1
celery==5.3.7
redis==5.2.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0